+++ SurfaceCaustics/Results_Final/Assets/Resources/Shaders/CausticColorShader_0.shader
@@ -62,9 +62,9 @@ Shader "Unlit/CausticColorShader"
                 return o;
             }
 
-            fixed4 frag(v2f i) : SV_Target
+            float4 frag(v2f i) : SV_Target
             {
-                return fixed4(1, 0, 0, 1);
+                return float4(1, 0, 0, 1);
             }
             ENDCG
         }
