+++ SurfaceCaustics/Results_Final/Assets/Resources/Shaders/CausticIntensityShader.shader
@@ -60,10 +60,10 @@ Shader "Unlit/CausticIntensityShader"
                 return o;
             }
 
-            fixed4 frag (v2f i) : SV_Target
+            float4 frag (v2f i) : SV_Target
             {
                 // sample the texture
-                fixed4 col = tex2D(_MainTex, i.uv);
+                float4 col = tex2D(_MainTex, i.uv);
                 float4 causticReceiverWorldPos = tex2D(_CausticMapTexture, i.uv);
                 float4 causticFlux = tex2D(_CausticFluxTexture, i.uv);
                 return col;
