+++ SurfaceCaustics/Results_Final/Assets/Resources/Shaders/CausticMapShader.shader
@@ -66,7 +66,7 @@ Shader "Unlit/CausticMapShader"
 
             float4 frag(v2f i) : SV_Target
             {
-                fixed4 col = tex2D(_MainTex, i.uv);
+                float4 col = tex2D(_MainTex, i.uv);
                 float isVisible = 0;
 
                 if (col.r != 0 || col.g != 0 && col.b != 0)
