+++ SurfaceCaustics/Results_Final/Assets/Resources/Shaders/CausticResultShader.shader
@@ -58,10 +58,10 @@ Shader "Unlit/CausticResultShader"
                 return o;
             }
 
-            fixed4 frag (v2f i) : SV_Target
+            float4 frag (v2f i) : SV_Target
             {
                 // sample the texture
-                fixed4 col = tex2D(_MainTex, i.uv);
+                float4 col = tex2D(_MainTex, i.uv);
                 return col;
             }
             ENDCG
