+++ SurfaceCaustics/Results_Final/Assets/Resources/Shaders/ConvertToGreenShader.shader
@@ -44,10 +44,10 @@ Shader "Unlit/ConvertToGreenShader"
                 return o;
             }
 
-            fixed4 frag (v2f i) : SV_Target
+            float4 frag (v2f i) : SV_Target
             {
                 // sample the texture
-                fixed4 col = tex2D(_MainTex, i.uv);
+                float4 col = tex2D(_MainTex, i.uv);
                 // apply fog
                 UNITY_APPLY_FOG(i.fogCoord, col);
                 return col;
