+++ SurfaceCaustics/Results_Final/Assets/Resources/Shaders/TestShader.shader
@@ -54,12 +54,12 @@ Shader "Unlit/TestShader"
                 return o;
             }
 
-            fixed4 frag(v2f i) : SV_Target
+            float4 frag(v2f i) : SV_Target
             {
                 // sample the texture
                 /*fixed4 col = tex2D(_MainTex, i.uv);
                 return col;*/
-                return fixed4(1, 1, 1, 1);
+                return float4(1, 1, 1, 1);
             }
             ENDCG
         }
